Add files via upload
The is the document template for each new document. The spreadsheet has three rows with student data. When the document is a Google Doc in Drive and the spreadsheet is a Google Sheet with the javascript running in the script editor, the "key**@" elements specified in the script are substituted using the copyDoc command for each row.